Created index.html
Copying in prototype created at  https://jsbin.com/xabucot 
Right now this doesn't allow custom PZ rides, it just does the xmas/hanuka ride that I created. 

An enhancement would be to alter the array or pull it in from a google sheet, or make it a tool that can take a list of intervals/timeslots and work based on that.